+++ pages/CustomerProfile/components/index.ts
@@ -15,6 +15,7 @@ export { default as EditCustomerModal } from './EditCustomerModal';
 export { default as EditRentalModal } from './EditRentalModal';
 export { default as InsightsSidebar } from './InsightsSidebar';
 export { default as RentalsSection } from './RentalsSection';
+export { default as ServicedExternalsSection } from './ServicedExternalsSection';
 export { default as RentForkliftModal } from './RentForkliftModal';
 export { default as ResultModal } from './ResultModal';
 export { default as ServiceHistory } from './ServiceHistory';